+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 ### Fixed
 - don't run in-memory shrinking if there is not enough memory available.
 - don't mark the project as in IDE sync mode when no tasks where called and no composite build detected.
+- remove invalid checks for composite mode.
 - improve and document composite builds detection.
 - properly quote and escape CLI arguments for external tool runner.
 - fix R8 external run for Ubuntu and macOS (non-Windows systems).